Fix gold implementations and eval tests for tcg_core API changes
- Add card_type/delta_species fields to all CardMeta initializers in 190 eval test files
- Fix Prompt initializers in gold implementations (missing fields)
- Fix delta_species.unwrap_or(false) on bool type
- Add find_pokemon/card_meta convenience methods to GameState

+++ gold/implementations/hp_025_muk.rs
@@ -14,7 +14,7 @@ pub fn apply_toxic_sludge(game: &mut GameState, muk_id: CardInstanceId) -> bool 
     // Put 1 damage counter on each active Pokemon
     for p in &mut game.players {
         if let Some(ref mut active) = p.active {
-            active.damage_counters += TOXIC_SLUDGE_COUNTERS;
+            active.damage_counters += TOXIC_SLUDGE_COUNTERS as u16;
         }
     }
     true
